+++ page-home.php
@@ -12,6 +12,8 @@ get_header();
 
 <?php get_template_part( 'template-parts/sections/hero' ); ?>
 
+<?php get_template_part( 'template-parts/sections/home-page-content' ); ?>
+
 <?php get_template_part( 'template-parts/sections/practice-areas-grid' ); ?>
 
 <?php get_template_part( 'template-parts/sections/cities-grid' ); ?>